Correct layers architecture
