--- SantasBag/SantasBag.Core/Domain/BaseEntity.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SantasBag.Core.Domain
-{
-    public abstract class BaseEntity
-    {
-        public Guid Id { get; set; }
-    }
-}
