--- SantasBag/SantasBag.Core/Domain/Contracts/RewardedDto.cs
@@ -1,10 +1,0 @@
-﻿namespace SantasBag.Core.Domain.Contracts
-{
-    public class RewardedDto //: BaseEntity
-    {
-        public required string Name { get; set; }
-        public string? Description { get; set; }
-        public int Cost { get; set; }
-        public Guid RoomId { get; set; }
-    }
-}
