+++ SantasBag/SantasBag.Core/Models/BaseEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace SantasBag.Core.Models;
+
+public abstract class BaseEntity
+{
+    public Guid Id { get; set; }
+}
