+++ SantasBag/SantasBag.sln
@@ -3,9 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34728.123
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SantasBag", "SantasBag\SantasBag.csproj", "{B3560934-4683-4586-9A8E-EF207D5406AB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SantasBag.WebHost", "SantasBag\SantasBag.WebHost.csproj", "{B3560934-4683-4586-9A8E-EF207D5406AB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SantasBag.Core", "SantasBag.Core\SantasBag.Core.csproj", "{CD1E3D68-3826-44AB-BE65-9B1004EEBEAB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SantasBag.Core", "SantasBag.Core\SantasBag.Core.csproj", "{CD1E3D68-3826-44AB-BE65-9B1004EEBEAB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SantasBug.DataAccess", "DataAccess\SantasBug.DataAccess.csproj", "{C8CA5AF7-ED67-4FA3-A420-374B76D2C942}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SantasBug.BusinessLogic", "SantasBug.BusinessLogic\SantasBug.BusinessLogic.csproj", "{10078E98-394D-4546-9435-392355A48351}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +25,14 @@ Global
 		{CD1E3D68-3826-44AB-BE65-9B1004EEBEAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CD1E3D68-3826-44AB-BE65-9B1004EEBEAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CD1E3D68-3826-44AB-BE65-9B1004EEBEAB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C8CA5AF7-ED67-4FA3-A420-374B76D2C942}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C8CA5AF7-ED67-4FA3-A420-374B76D2C942}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C8CA5AF7-ED67-4FA3-A420-374B76D2C942}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C8CA5AF7-ED67-4FA3-A420-374B76D2C942}.Release|Any CPU.Build.0 = Release|Any CPU
+		{10078E98-394D-4546-9435-392355A48351}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10078E98-394D-4546-9435-392355A48351}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10078E98-394D-4546-9435-392355A48351}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10078E98-394D-4546-9435-392355A48351}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
