+++ SantasBag/SantasBag/Contracts/PresentedRewards.cs
@@ -1,0 +1,9 @@
+﻿using SantasBag.Core.Models;
+
+namespace SantasBag.WebHost.Contracts;
+
+public class PresentedRewards: BaseEntity
+{
+    public required RewardedDto reward { get; set; }
+    public required Guid buyer { get; set; }
+}
