+++ SantasBag/SantasBag/Contracts/RewardedDto.cs
@@ -1,0 +1,9 @@
+﻿namespace SantasBag.WebHost.Contracts;
+
+public class RewardedDto //: BaseEntity
+{
+    public required string Name { get; set; }
+    public string? Description { get; set; }
+    public int Cost { get; set; }
+    public Guid RoomId { get; set; }
+}
