+++ SantasBag/SantasBag/SantasBag.WebHost.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SantasBag.Core\SantasBag.Core.csproj" />
+  </ItemGroup>
+
 </Project>
